+++ backend/uploader/uploadBanner.go
@@ -8,13 +8,17 @@ import (
 
 //UploadBanner uploads an banner picture while returning the image link and error by first replacing the old image and then uploading a new image
 func (uploader *Uploader) UploadBanner(ctx context.Context, req *uploaderpb.UploadBannerRequest) (*uploaderpb.UploadBannerResponse, error) {
-	err := uploader.Delete(ctx, req.OldImageLink)
-
-	if err != nil {
-		return nil, err
+	if req.Base64 == "" {
+		return nil, errBase64Empty
 	}
-
 	newLink, err := uploader.Upload(ctx, req.Base64, req.UUID)
+
+	if req.OldImageLink != "" {
+		err := uploader.Delete(ctx, req.OldImageLink)
+		if err != nil {
+			return nil, err
+		}
+	}
 
 	return &uploaderpb.UploadBannerResponse{NewLink: newLink}, err
 }
